Consolidate sanity dirs
