+++ src/lib/sanity/client.ts
@@ -1,6 +1,6 @@
 import { createClient } from 'next-sanity';
 
-import { apiVersion, dataset, projectId, useCdn } from '../../../sanity/env';
+import { apiVersion, dataset, projectId, useCdn } from './env';
 
 export const client = createClient({
   apiVersion,
